Menu stuff
Can now pause the game, save the game, and exit back to level select. You cannot load the data yet.

Next big step, an actual Main Menu with a title. I think for this main menu it will change with planets added, so first iteration it is all Virellia, second iteration it will be a level composed of both Virellia and Planet2 elements, and etc until maybe 5 planets?